Replace grep in boolean context with any()
This is not only theoretically faster than the grep solution, but also
the intention of the code is a bit clearer; one also gets to see that a
year is a number and not a string, which means we no longer need a
match.  This was picked up by kritika.io.