Check that regexes are successful before using captures
This is according to the RegularExpressions::ProhibitCaptureWithoutTest
Perl::Critic policy.  It's possible that a regex fails and hence the
capture value used in the following code uses an old, incorrect value of
the capture variable.  By checking for the success of the regex one can
ensure that the capture is that which one expects.  This issue was found
by kritika.io.