Specify coefficients order for Math::Polynomial::Solve
This module's behaviour will change as part of an upcoming version 3.00
release and in order to protect currently-existing code, the
single default export from the module (`coefficients`) needs to have its
ordering specified explicitly.  When running the tests, the module gave
this warning:

```
Your coefficient order is in a default state, which will change by version 3.00.

Please put
        coefficients order => 'descending';
at the beginning of file t/04-ice-extent-poly-fit.t to make
sure your function parameters will be in the correct order when the
default order changes.

See the README file and the Math::Polynomial::Solve documentation for
more information.
```

Therefore, I explicitly import `coefficients` and set its order to
`'descending'` as recommended by the module author.