Use perlimports to restrict imports in main program
The program `perlimports` works out exactly what one is using in one's
code, so that the imports can be restricted to exactly those things that
are needed and no more.  This can help remove conflicts between packages
and we avoid importing things we don't need, which is cleaner and less
likely to cause unwanted issues.